+++ bot/cogs/admin.py
@@ -293,6 +293,11 @@ class AdminCog(commands.Cog, name='admin'):
                 await loop.run_in_executor(None, db.get_all_elements, classes.Base, "static_bases")
                 await disp.BOT_RELOAD.send(ctx, "Bases")
                 return
+            if arg == "config":
+                await loop.run_in_executor(None, cfg.get_config, cfg.LAUNCH_STR)
+                await roles.update_rule_msg()
+                await disp.BOT_RELOAD.send(ctx, "Config")
+                return
         await disp.WRONG_USAGE.send(ctx, ctx.command.name)
 
     @commands.command(aliases=['rm'])