+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added doc folder
 - Accounts are now added to player objects before validation
 - Bases can now be picked with pick command
+- Config can now be reloaded with reload command
 
 # v3.2:
 - Added locking mechanic on reaction handler to avoid errors due to players spamming the reactions